src/requirements.txt
flask==2.0.1
neo4j==5.3.0
requests==2.31.0
Pillow==10.0.0
PyMuPDF==1.23.0
python-docx==0.8.11
Werkzeug==2.0.1 